Check for mirrorlist age when optimizing
Given mirrors are servers on the internet, bound to their owner's wishes, they are likely to suffer outages or even death. Thus, leading to an "invalid" list of mirrors, which you might note if the dead one is up on your list. That leads to annoying logs and installation slowdowns once pacman hits every server until it finds the package. Therefore, you might get a slower second run after a prior update to the mirror list.

To mitigate that a little, I added a check for the age of mirrors. If the mirrorlist file is older than seven days, we'll update it whether or not the mirrorlist was optimized.